+++ common/src/main/java/com/tristankechlo/livingthings/entity/AncientBlazeEntity.java
@@ -18,12 +18,12 @@ import net.minecraft.server.level.ServerBossEvent;
 import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.sounds.SoundEvent;
 import net.minecraft.sounds.SoundSource;
+import net.minecraft.tags.DamageTypeTags;
 import net.minecraft.tags.FluidTags;
 import net.minecraft.world.BossEvent;
 import net.minecraft.world.Difficulty;
 import net.minecraft.world.DifficultyInstance;
 import net.minecraft.world.damagesource.DamageSource;
-import net.minecraft.world.damagesource.DamageTypes;
 import net.minecraft.world.entity.*;
 import net.minecraft.world.entity.ai.attributes.AttributeSupplier;
 import net.minecraft.world.entity.ai.attributes.Attributes;
@@ -175,8 +175,8 @@ public class AncientBlazeEntity extends Monster implements PowerableMob, RangedA
 
     @Override
     public boolean hurt(DamageSource source, float amount) {
-        // dont get damaged while charging up
-        if (this.getInvulnerableTime() > 0 && source.typeHolder().is(DamageTypes.OUT_OF_WORLD.location())) {
+        // don't get damaged while charging up
+        if (this.getInvulnerableTime() > 0 && !source.is(DamageTypeTags.BYPASSES_INVULNERABILITY)) {
             return false;
             // catch large fireballs
         } else if (source.getDirectEntity() instanceof LargeFireball && source.getEntity() instanceof Player) {
@@ -189,10 +189,9 @@ public class AncientBlazeEntity extends Monster implements PowerableMob, RangedA
             // random chance for arrows, tridents,.. to be blocked
         } else if (source.isIndirect()) {
             return this.random.nextInt(4) != 0 && super.hurt(source, amount);
-        } else {
-            // normal damage handling
-            return super.hurt(source, amount);
         }
+        // normal damage handling
+        return super.hurt(source, amount);
     }
 
     @Override